Add BuildToolLayout for testing and run DAP suites with sbt sever
The workspace layout of a test depends on the build tool we want to use.
By using a BuildToolLayout abstraction we can define tests that are not
specific to a build tool and then run the tests on many build tools.

We use the BuildToolLayout to abstract over the DAP suites and run them
with sbt server as the build server.
