+++ tests/slow/src/test/scala/tests/sbt/SbtBreakpointDapSuite.scala
@@ -1,0 +1,12 @@
+package tests.sbt
+
+import tests.SbtBuildLayout
+import tests.SbtServerInitializer
+import tests.debug.BaseBreakpointDapSuite
+
+class SbtBreakpointDapSuite
+    extends BaseBreakpointDapSuite(
+      "sbt-debug-breakpoint",
+      SbtServerInitializer,
+      SbtBuildLayout
+    )
