+++ tests/slow/src/test/scala/tests/sbt/SbtCrossDebugSuite.scala
@@ -1,0 +1,12 @@
+package tests.sbt
+
+import tests.SbtBuildLayout
+import tests.SbtServerInitializer
+import tests.feature.BaseCrossDebugSuite
+
+class SbtCrossDebugSuite
+    extends BaseCrossDebugSuite(
+      "sbt-cross-debug",
+      SbtServerInitializer,
+      SbtBuildLayout
+    )
