+++ tests/slow/src/test/scala/tests/sbt/SbtStepDapSuite.scala
@@ -1,0 +1,12 @@
+package tests.sbt
+
+import tests.SbtBuildLayout
+import tests.SbtServerInitializer
+import tests.debug.BaseStepDapSuite
+
+class SbtStepDapSuite
+    extends BaseStepDapSuite(
+      "sbt-debug-step",
+      SbtServerInitializer,
+      SbtBuildLayout
+    )
