+++ tests/unit/src/main/scala/tests/BuildServerLayout.scala
@@ -1,0 +1,43 @@
+package tests
+
+import scala.meta.internal.metals.{BuildInfo => V}
+
+sealed trait BuildToolLayout {
+  def apply(sourceLayout: String, scalaVersion: String): String
+}
+
+object QuickBuildLayout extends BuildToolLayout {
+  override def apply(
+      sourceLayout: String,
+      scalaVersion: String
+  ): String = {
+    s"""|/metals.json
+        |{ 
+        |  "a": {"scalaVersion": "$scalaVersion"},  "b": {"scalaVersion": "$scalaVersion"} 
+        |}
+        |$sourceLayout
+        |""".stripMargin
+  }
+}
+
+object SbtBuildLayout extends BuildToolLayout {
+  val commonSbtSettings: String =
+    """|import scala.concurrent.duration._
+       |Global / serverIdleTimeout := Some(1 minute)
+       |""".stripMargin
+
+  override def apply(
+      sourceLayout: String,
+      scalaVersion: String
+  ): String = {
+    s"""|/project/build.properties
+        |sbt.version=${V.sbtVersion}
+        |/build.sbt
+        |$commonSbtSettings
+        |ThisBuild / scalaVersion := "$scalaVersion"
+        |val a = project.in(file("a"))
+        |val b = project.in(file("b"))
+        |$sourceLayout
+        |""".stripMargin
+  }
+}
